Mv celery monitoring to each own package dir
